+++ mpixdaq/_version_info.py
@@ -9,7 +9,7 @@
 major = 1
 minor = 0
 revision = 1
-suffix = "rc1"  # for suffixes such as 'rc' or 'beta' or 'alpha'
+suffix = "rc2"  # for suffixes such as 'rc' or 'beta' or 'alpha'
 
 
 def _get_version_tuple():